Move test to dask array module
